+++ crates/fuzzing/src/generators.rs
@@ -15,7 +15,7 @@ mod config;
 mod instance_allocation_strategy;
 mod instance_limits;
 mod memory;
-mod module_config;
+mod module;
 mod single_inst_module;
 mod spec_test;
 mod stacks;
@@ -26,7 +26,7 @@ pub use config::{Config, WasmtimeConfig};
 pub use instance_allocation_strategy::InstanceAllocationStrategy;
 pub use instance_limits::InstanceLimits;
 pub use memory::{MemoryConfig, NormalMemoryConfig, UnalignedMemory, UnalignedMemoryCreator};
-pub use module_config::ModuleConfig;
+pub use module::{Module, ModuleConfig};
 pub use single_inst_module::SingleInstModule;
 pub use spec_test::SpecTest;
 pub use stacks::Stacks;
